Renamed Logger.py to logger.py for consistency
